fix the previous sorting bugfix, add some explanation for future me
The old bug was that I misunderstood my own logic a few months after
writing. Hence the notes now.
The problem was just that I didn't realize that `i == df.high` and any
other case before that where we want to sort (`last != cur`) are
different in terms of the last idx to be sorted.