use custom clone even for non ARC due to laser ptr len tensors
Deepcopying laser tensors using ptr + len pairs is a really bad
idea. Only the pointer will be copied. This means that the copy of the
old tensor will be invalid the moment the original goes out of scope!
Aside from that a deep copy is never deep here. The original data is
mutated!

Causes random segfaults. :)