[formula] simplify Preface to work with NimNode directly
Adds a `node: NimNode` field to `Assign`, which is used to identify
the exact `NimNode` that we need to replace.
This avoids having to rely on a lot of stringification and using only
the column names. This is really important when dealing with the same
column being accessed but requiring different replacements (e.g. once
indices and once full column).