make `filter` work w/ groupd DFs & scalar formulas
Grouped + scalar formulas can be used to filter on full DF subgroups.

This also introduces `FormulaMismatchError` a catchable exception that
is thrown in case a procedure receives a formula that does not match
the expected kinds.

It would be really neat if we could differentiate at CT.

We could in theory make =FormulaNode= generic and differentiate
between the different kinds + different return types...