fix CSV parsing of empty fields in first row if not float col
If a column missed elements in the first row, but then wasn't a float
column, the result was that the column was first interpreted to be
float, but then realized it doesn't contain float, so the result was
an object column, when e.g. an int column would have been fine.