add convenience `Value` comparison ops valid within f{} macro
This adds the most common boolean operators that one might use on an
object column within a formula. To avoid introducing these everywhere,
we emit the templates within the scope of the formula closure.

The user can call the template manually to define them in a local
scope of course.