impr. constant col support by less const->full transf, use native types
This avoids a large number (but not all unfortunately) of the
transformations that were previously required when doing operations
with constant columns. Now many more are supported without any change.

If a constant column is transformed to a full column, at least the
native type of the underlying constant `Value` is now chosen
for *much* better performance if the resulting column is used in computations.