special case `ifStmt` in type determination logic
It is not sensible to use the if conditionals as a way to determine if
something is a *bool*, as the expression of course *has* to be a
bool. These branches are of course still used to determine associated
types from e.g. `foo < 1`, which tells us `foo` is probably an integer.