add logic for formulas with symbols that are resolved to `None`
`None` symbols apparently appear if the input is using a procedure on
an argument with column reference (such that the argument doesn't make
any sense in terms of the types for the compiler). In that case a
symbol when pushed to the typed stage might end up being `None`.

This handles these in an ugly way by using `when compiles` :(