allow gen of `for` loop in reducing `<<` formula if col idx used
Before this reducing formula had to use a reducing procedure on a
tensor. Now, if an index is accessed in the reducing formula, we
generate a for loop, which accumulates the yielded values.

Further support outside of `res += ...` will be added in the future
once I come up with a nice way to write that code (thinking about just
forcing the user to be explicit in this case, i.e. if `<<` is used in
combination with an index access to a column, CT error saying:

Please rewrite your formula using:

`res += / -= / ... <formula body`

Maybe. I.e. tell it explicitly what to do with the accumulated values.

In addition this commit lays the foundation for the support of
transformed nodes, i.e. letting the user transform a full tensor
before the loop.