resurrect full `formula: ` syntax with a preface and loop body
For more complex statements this kind of formula is very useful.
Example:

```nim
 let fn = formula:
     preface:
       lagIdx in lag(df["Time / s", float])
     loop:
       "sumTime" << (block:
                  if idx("B / T") > 1.0:
                    idx("Time / s") - lagIdx
                  else:
                         0.0)
 echo df.summarize(fn).pretty(precision = 10)
```

where we apply the `lag` procedure to the single data frame column and
use it individual in a reducing formula.

NOTE: this comes at a cost: namely these formulas *cannot* be used as
arguments to functions directly, due to a parser limitation.

The `preface` defines variables and their corresponding tensors (where
one needs to use the `df` name for the input DF). These are
placed *above* the loop and the LHS is used as the variable name in
the body of the for loop.

NOTE: this might still be broken for the case of non memcopyable
types, i.e. when we generate a for loop with explicit index access
instead of a `forEach`.