fixup `toTab` & related to make `toDf` work equivalently to `seqsToDf`
This is something that I should have done much earlier.. :)

The typed overloads for `toDf` broke the `toDf` template that was
supposed to be equivalent to `seqsToDf`.

However, I never wanted to break code that used `toDf` for things like
an `OrderedTable` or as a no-op from an old `readCsv` call. This
finally, fixes up the `toTab` macro to handle these things correctly.