replace assertion check by exception w/ message, add test
Reported by ezquerra on the #science channel.

The assertion was naive, as there was nothing preventing the argument
to be a non reducing formula. I was under the wrong impression that
the formula macro itself would catch all cases where wrong formulas
are handed, but that's obviously not the case.

In theory we could have separate `Formula` types for the different
`FormulaKinds`. Problem is that then we are back in land of generics hell with
different possibilities if things are mixed in one procedure etc. :(