[io] ignore header if `colNames` are given
If the user gives custom column names to the `readCsv` procedure, we
skip the `parseHeaderCol` call. It can only cause problems (e.g. the
user wishes to ignore the header as it's not conforming to the numbers
of columns according to the separator), and does not bring any
advantage.

This _does_ mean however, that we are "off" by one `skipLine` now
compared to before. Now `colNames` _must_ be combined with a single
`skipLine`! We cannot set it to 1 automatically, because the user may
want to use `colNames` because there is no header at all!