[io] add support for reading CSV files from http and https URLs
Make it possible to specify a web URL as the fname argument of readCsv. If fname starts with http or https the file _contents_ will be automatically downloaded before being parsed. Note that no catching is performed (i.e. the file contents are downloaded every time) and that -d:ssl is needed in order to download from https URLs.