improve type deduction logic for return types
We now not only look at the deduced types for the return value, but
also at the heuristic types. The latter though with a lower
weighting. The idea is that if the heuristic is `bool` we probably
want to keep that information (as it will likely come from something
like `in` or similar, which due to being a template can only be
deduced from our heuristic magic). But if the heuristic says float,
but the deduced one says int we do *not* want to override it.

Further we added logic to determine the *last* type of the AST in
`determineTypesImpl`. Last in the sense of being the likely
result (i.e. last child of `nnkDotExpr`). This feeds into the deduced
type info for the logic mentioned above.