[tests] update test cases for {.error.} pragma regressions
Without the previous commit of checking for the error pragma the
original tests did now not pass. We were generating code `Tensor +
Scalar` in the previous version for this. But that was never actually
intended. Technically it was simply allowed and 'correct' given our
logic and available tools in arraymancer.